refactor(portal): move authenticator into package
